Rewrite a lot of code to caculate what's provisioned and how much space that will take on disk.
This commit adds a bunch of new methods and refactors some stuff.
* A new function to get all pool names.
* A method to get pool size. It first runs rbd du, and if the there's no rbd
disk use for that pool, then it will attempt to find rgw/cephfs usage from
ceph df.
* Method to get replication factor. For instance, if a pool is triple replicated, then
the factor is simply 3. For erasure coded pools the math is based on how many parity drives we have. We multiply out pool size with this factor to caclulate the space required on disk.
* A method to get which pool belongs to what root, since we are interested in breakdown
per ceph root (not ceph pool).